feat(spaceLayout): setting sidebar in mobile/tablet mode
