+++ app/spaces/[spaceid]/edit/components/backHeader.tsx
@@ -1,0 +1,31 @@
+import { CaretLeft } from '@phosphor-icons/react';
+import { useRouter } from 'next/navigation';
+
+export interface IBackHeaderProps {
+  spaceId: string;
+}
+
+const BackHeader = ({ spaceId }: IBackHeaderProps) => {
+  const router = useRouter();
+
+  const handleBack = () => {
+    router.push(`/spaces/${spaceId}`);
+  };
+
+  return (
+    <div className="flex items-center gap-[10px] w-full ">
+      {/* 返回按钮 */}
+      <button
+        className="h-[34px] bg-[#363636] hover:bg-[#424242] rounded-lg px-3.5 flex items-center gap-[5px]"
+        onClick={handleBack}
+      >
+        <CaretLeft size={18} weight="light" format={'Stroke'} className="text-white" />
+        <span className="text-white text-[13px] font-medium">Back</span>
+      </button>
+
+      <span className="text-white text-[14px]">Space Settings</span>
+    </div>
+  );
+};
+
+export default BackHeader;
